requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
pyodbc>=5.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
